+++ src/main/java/com/smartCity/Web/Model/MarketRate.java
@@ -7,7 +7,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import lombok.Data;
 
 
 
@@ -16,7 +15,7 @@ import lombok.Data;
 public class MarketRate {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 	private String productName;
 	private Double price;
 	private String unit;
@@ -32,11 +31,11 @@ public class MarketRate {
 		this.priceDate = LocalDate.now();
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
