+++ src/main/java/com/smartCity/Web/Repository/MarketRateRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.smartCity.Web.Model.MarketRate;
 
 @Repository
-public interface MarketRateRepository extends JpaRepository<MarketRate, Long> {
+public interface MarketRateRepository extends JpaRepository<MarketRate, Integer> {
 }
